Add Close button and refactor common components file structure
