+++ src/features/common/components/Container.tsx
@@ -1,6 +1,6 @@
 import { type PropsWithChildren } from "react";
 
-export default function Container({ children }: PropsWithChildren) {
+export function Container({ children }: PropsWithChildren) {
   return (
     <div className="m-auto box-border flex h-full w-full flex-col p-2 sm:max-w-2xl">
       {children}
