+++ src/features/common/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./Button";
+export * from "./Container";
+export * from "./Loader";
+export * from "./Modal";
