+++ src/utils/schemas/food.ts
@@ -1,7 +1,9 @@
 import { z } from "zod";
 import { FoodTypes } from "@prisma/client";
 
-export const getFoods = z.object({ ubicationId: z.string().min(1) });
+export const getFoodsFromUbication = z.object({
+  ubicationId: z.string().min(1),
+});
 
 export const createFood = z.object({
   ubicationId: z.string().min(1),
